bedre plassering av metoder. Default Mapper for typegenerering
