+++ web/src/main/java/no/nav/foreldrepenger/web/app/jackson/JacksonJsonConfig.java
@@ -1,24 +1,11 @@
 package no.nav.foreldrepenger.web.app.jackson;
 
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 import jakarta.ws.rs.ext.ContextResolver;
 import jakarta.ws.rs.ext.Provider;
 
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 
-import no.nav.foreldrepenger.domene.opptjening.dto.AvklarAktivitetsPerioderDto;
-import no.nav.foreldrepenger.domene.person.verge.dto.AvklarVergeDto;
-import no.nav.foreldrepenger.domene.rest.dto.VurderFaktaOmBeregningDto;
-import no.nav.foreldrepenger.familiehendelse.aksjonspunkt.omsorgsovertakelse.dto.VurderOmsorgsovertakelseVilkårAksjonspunktDto;
-import no.nav.foreldrepenger.web.app.IndexClasses;
 import no.nav.foreldrepenger.web.app.tjenester.RestImplementationClasses;
 import no.nav.vedtak.mapper.json.DefaultJsonMapper;
 
@@ -28,18 +15,8 @@ public class JacksonJsonConfig implements ContextResolver<ObjectMapper> {
     private static final JsonMapper JSON_MAPPER = createObjectMapper();
 
     private static synchronized JsonMapper createObjectMapper() {
-        var typeNameClasses = allJsonTypeNameClasses();
+        var typeNameClasses = RestImplementationClasses.allJsonTypeNameClasses();
         return DefaultJsonMapper.getJsonMapper().rebuild().registerSubtypes(typeNameClasses).build();
-    }
-
-    /**
-     * Scan subtyper dynamisk fra WAR slik at superklasse slipper å
-     * deklarere @JsonSubtypes.
-     */
-    public static List<Class<?>> getJsonTypeNameClasses(URI classLocation) {
-        IndexClasses indexClasses;
-        indexClasses = IndexClasses.getIndexFor(classLocation);
-        return indexClasses.getClassesWithAnnotation(JsonTypeName.class);
     }
 
     @Override
@@ -47,25 +24,4 @@ public class JacksonJsonConfig implements ContextResolver<ObjectMapper> {
         return JSON_MAPPER;
     }
 
-    public static Set<Class<?>> allJsonTypeNameClasses() {
-        final var scanClasses = new LinkedHashSet<>(RestImplementationClasses.getImplementationClasses());
-
-        scanClasses.add(AvklarAktivitetsPerioderDto.class);
-        scanClasses.add(VurderFaktaOmBeregningDto.class);
-        scanClasses.add(VurderOmsorgsovertakelseVilkårAksjonspunktDto.class);
-        scanClasses.add(AvklarVergeDto.class);
-
-        return scanClasses.stream()
-            .map(c -> {
-                try {
-                    return c.getProtectionDomain().getCodeSource().getLocation().toURI();
-                } catch (URISyntaxException e) {
-                    throw new IllegalArgumentException("Ikke en URI for klasse: " + c, e);
-                }
-            })
-            .distinct()
-            .flatMap(uri -> getJsonTypeNameClasses(uri).stream())
-            .collect(Collectors.toUnmodifiableSet());
-
-    }
 }
